requirements.txt
python-dotenv==1.0.0
psycopg2-binary==2.9.9
google-generativeai==0.3.2
PyPDF2==3.0.1
python-docx==1.1.0
pgvector==0.2.4
